requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
boto3==1.38.3
botocore==1.38.3
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
packaging==25.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
whitenoise==6.9.0
